dev_requirements.txt
###### production ######
django==1.11.7  # framework
python-memcached  # session cache
validate_email
google-auth
requests
pillow
django-cachalot
django-memcache-status
facebook-sdk
pytesseract
